Reimplemented "experimental.retroTerminalEffect" and toggle commands
After review it was clear for backwards compatibility reasons we need
preserve "experimental.retroTerminalEffect".

In addition, the command "toggleRetroEffect" now turns the effects on
and off again together with the new "toggleTerminalEffects".

"toggleRetroEffect" is kept for backwars compatibility reasons. Slight
change in semantics in that it requires a terminal effect to be
configured in order to be able toggle it on and off.