Added experimental.pixelShaderImagePath
This is to support customer shaders with custom textures.

I was thinking of exposing the background image to the shader but
that felt complicated after looking into it.

Instead I load the texture using WIC into a shader resource view.
When binding shader resources I test for presence of customer texture
and bind it to register t1.

The image loading code was found here:

https://learn.microsoft.com/en-us/windows/win32/direct3d11/overviews-direct3d-11-resources-textures-how-to